[Silver III] Title: 주유소, Time: 388 ms, Memory: 36460 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,19 +8,19 @@ class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N = Integer.parseInt(br.readLine());
 
-		int[] dis = new int[N - 1];
+		long[] dis = new long[N - 1];
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < N - 1; i++) {
 			dis[i] = Integer.parseInt(st.nextToken());
 		}
 
-		int[] cost = new int[N - 1];
+		long[] cost = new long[N - 1];
 		st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < N - 1; i++) {
-			cost[i] = Integer.parseInt(st.nextToken());
+			cost[i] = Long.parseLong(st.nextToken());
 		}
 
-		int minCost = 1000000000;
+		long minCost = 1000000000;
 		long ans = 0;
 		for (int i = 0; i < N - 1; i++) {
 			if (minCost > cost[i]) {
